requirements.txt
amqp==2.1.4
anyjson==0.3.3
argcomplete==1.7.0
base58==0.2.4
billiard==3.5.0.2
boto3==1.4.2
botocore==1.4.87
celery==4.0.2
celery-redbeat==0.11.0
click==6.6
dj-database-url==0.4.1
Django==1.10.4
django-activity-stream==0.6.3
django-allauth==0.29.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-cors-headers==2.1.0
django-jet==1.0.6
django-jsonfield==1.0.1
django-model-utils==2.6.1
django-tastypie==0.13.3
docutils==0.13.1
futures==3.0.5
google-api-python-client==1.4.1
gunicorn==19.6.0
hjson==2.0.2
httplib2==0.10.3
jmespath==0.9.0
kappa==0.6.0
kombu==4.0.2
lambda-packages==0.11.0
lxml==3.7.1
oauth2client==4.1.2
oauthlib==2.0.1
placebo==0.8.1
psycopg2==2.6.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
python-dateutil==2.6.0
python-mimeparse==1.6.0
python-openid==2.2.5
python-slugify==1.2.1
pytz==2016.10
PyYAML==3.12
redis==2.10.5
requests==2.12.4
requests-oauthlib==0.7.0
rsa==3.4.2
s3transfer==0.1.10
simplejson==3.10.0
six==1.10.0
toml==0.9.2
tqdm==4.10.0
troposphere==1.9.0
Unidecode==0.4.19
uritemplate==3.0.0
vine==1.1.3
Werkzeug==0.11.11
whitenoise==3.2.2
wsgi-request-logger==0.4.6
zappa==0.32.1
